+++ compiler-core/src/type_.rs
@@ -25,6 +25,7 @@ use crate::{
     },
     bit_string,
     build::{Origin, Target},
+    uid::UniqueIdGenerator,
 };
 use expression::*;
 
@@ -147,7 +148,7 @@ impl Type {
         module: &[String],
         name: &str,
         arity: usize,
-        environment: &mut Environment<'_, '_>,
+        environment: &mut Environment<'_>,
     ) -> Option<Vec<Arc<Self>>> {
         match self {
             Self::App {
@@ -336,9 +337,9 @@ pub enum PatternConstructor {
 
 #[derive(Debug, Clone, PartialEq)]
 pub enum TypeVar {
-    Unbound { id: usize },
+    Unbound { id: u64 },
     Link { type_: Arc<Type> },
-    Generic { id: usize },
+    Generic { id: u64 },
 }
 
 impl TypeVar {
@@ -429,7 +430,7 @@ impl ValueConstructor {
 ///
 pub fn infer_module(
     target: Target,
-    uid: &mut usize,
+    ids: &UniqueIdGenerator,
     mut module: UntypedModule,
     origin: Origin,
     package: &str,
@@ -438,7 +439,7 @@ pub fn infer_module(
 ) -> Result<TypedModule, Error> {
     let name = module.name.clone();
     let documentation = std::mem::take(&mut module.documentation);
-    let mut environment = Environment::new(uid, &name, modules, warnings);
+    let mut environment = Environment::new(ids.clone(), &name, modules, warnings);
     validate_module_name(&name)?;
 
     let mut type_names = HashMap::with_capacity(module.statements.len());
@@ -606,7 +607,7 @@ fn register_values<'a>(
     module_name: &[String],
     hydrators: &mut HashMap<String, Hydrator>,
     names: &mut HashMap<&'a str, &'a SrcSpan>,
-    environment: &mut Environment<'_, '_>,
+    environment: &mut Environment<'_>,
 ) -> Result<(), Error> {
     match s {
         Statement::Fn {
@@ -848,7 +849,7 @@ fn register_values<'a>(
 fn generalise_statement(
     s: TypedStatement,
     module_name: &[String],
-    environment: &mut Environment<'_, '_>,
+    environment: &mut Environment<'_>,
 ) -> TypedStatement {
     match s {
         Statement::Fn {
@@ -918,7 +919,7 @@ fn infer_statement(
     s: UntypedStatement,
     module_name: &[String],
     hydrators: &mut HashMap<String, Hydrator>,
-    environment: &mut Environment<'_, '_>,
+    environment: &mut Environment<'_>,
 ) -> Result<TypedStatement, Error> {
     match s {
         Statement::Fn {
@@ -1330,7 +1331,7 @@ fn assert_no_labelled_arguments<A>(args: &[CallArg<A>]) -> Result<(), Error> {
 /// could cause naively-implemented type checking to diverge.
 /// While traversing the type tree.
 ///
-fn unify_unbound_type(typ: Arc<Type>, own_id: usize) -> Result<(), UnifyError> {
+fn unify_unbound_type(typ: Arc<Type>, own_id: u64) -> Result<(), UnifyError> {
     if let Type::Var { type_: typ } = typ.deref() {
         let new_value = match typ.borrow().deref() {
             TypeVar::Link { type_: typ, .. } => return unify_unbound_type(typ.clone(), own_id),
@@ -1381,7 +1382,7 @@ fn unify_unbound_type(typ: Arc<Type>, own_id: usize) -> Result<(), UnifyError> {
 fn match_fun_type(
     typ: Arc<Type>,
     arity: usize,
-    environment: &mut Environment<'_, '_>,
+    environment: &mut Environment<'_>,
 ) -> Result<(Vec<Arc<Type>>, Arc<Type>), MatchFunTypeError> {
     if let Type::Var { type_: typ } = typ.deref() {
         let new_value = match typ.borrow().deref() {
@@ -1469,7 +1470,7 @@ fn make_type_vars(
     args: &[String],
     location: &SrcSpan,
     hydrator: &mut Hydrator,
-    environment: &mut Environment<'_, '_>,
+    environment: &mut Environment<'_>,
 ) -> Result<Vec<Arc<Type>>, Error> {
     args.iter()
         .map(|arg| TypeAst::Var {
@@ -1483,7 +1484,7 @@ fn make_type_vars(
 fn custom_type_accessors<A>(
     constructors: &[RecordConstructor<A>],
     hydrator: &mut Hydrator,
-    environment: &mut Environment<'_, '_>,
+    environment: &mut Environment<'_>,
 ) -> Result<Option<HashMap<String, RecordAccessor>>, Error> {
     // Get the constructor for this custom type.
     let args = match constructors {
@@ -1517,7 +1518,7 @@ pub fn register_types<'a>(
     module: &[String],
     hydrators: &mut HashMap<String, Hydrator>,
     names: &mut HashMap<&'a str, &'a SrcSpan>,
-    environment: &mut Environment<'_, '_>,
+    environment: &mut Environment<'_>,
 ) -> Result<(), Error> {
     match statement {
         Statement::ExternalType {
@@ -1645,7 +1646,7 @@ pub fn register_types<'a>(
 
 pub fn register_import(
     s: &UntypedStatement,
-    environment: &mut Environment<'_, '_>,
+    environment: &mut Environment<'_>,
 ) -> Result<(), Error> {
     match s {
         Statement::Import {