+++ compiler-core/src/type_/tests/errors.rs
@@ -5,15 +5,15 @@ macro_rules! assert_module_error {
         let (mut ast, _) = crate::parse::parse_module($src).expect("syntax error");
         ast.name = vec!["my_module".to_string()];
         let mut modules = HashMap::new();
-        let mut uid = 0;
+        let ids = UniqueIdGenerator::new();
         // DUPE: preludeinsertion
         // TODO: Currently we do this here and also in the tests. It would be better
         // to have one place where we create all this required state for use in each
         // place.
-        let _ = modules.insert("gleam".to_string(), build_prelude(&mut uid));
+        let _ = modules.insert("gleam".to_string(), build_prelude(&ids));
         let ast = infer_module(
             Target::Erlang,
-            &mut uid,
+            &ids,
             ast,
             Origin::Src,
             "thepackage",
@@ -28,15 +28,15 @@ macro_rules! assert_module_error {
         use std::path::PathBuf;
         let (ast, _) = crate::parse::parse_module($src).expect("syntax error");
         let mut modules = HashMap::new();
-        let mut uid = 0;
+        let ids = UniqueIdGenerator::new();
         // DUPE: preludeinsertion
         // TODO: Currently we do this here and also in the tests. It would be better
         // to have one place where we create all this required state for use in each
         // place.
-        let _ = modules.insert("gleam".to_string(), build_prelude(&mut uid));
+        let _ = modules.insert("gleam".to_string(), build_prelude(&ids));
         let error = infer_module(
             Target::Erlang,
-            &mut uid,
+            &ids,
             ast,
             Origin::Src,
             "thepackage",
@@ -57,16 +57,16 @@ macro_rules! assert_module_error {
 macro_rules! assert_error {
     ($src:expr, $error:expr $(,)?) => {
         let ast = crate::parse::parse_expression_sequence($src).expect("syntax error");
-        let mut uid = 0;
+        let ids = UniqueIdGenerator::new();
         let mut modules = HashMap::new();
         // DUPE: preludeinsertion
         // TODO: Currently we do this here and also in the tests. It would be better
         // to have one place where we create all this required state for use in each
         // place.
-        let _ = modules.insert("gleam".to_string(), build_prelude(&mut uid));
+        let _ = modules.insert("gleam".to_string(), build_prelude(&ids));
         println!("new assert_error test: {}", modules.len());
         let result = ExprTyper::new(&mut Environment::new(
-            &mut uid,
+            ids,
             &["somemod".to_string()],
             &modules,
             &mut vec![],
@@ -79,16 +79,16 @@ macro_rules! assert_error {
     ($src:expr) => {
         use std::path::PathBuf;
         let ast = crate::parse::parse_expression_sequence($src).expect("syntax error");
-        let mut uid = 0;
+        let ids = UniqueIdGenerator::new();
         let mut modules = HashMap::new();
         // DUPE: preludeinsertion
         // TODO: Currently we do this here and also in the tests. It would be better
         // to have one place where we create all this required state for use in each
         // place.
-        let _ = modules.insert("gleam".to_string(), build_prelude(&mut uid));
+        let _ = modules.insert("gleam".to_string(), build_prelude(&ids));
         println!("new assert_error test: {}", modules.len());
         let error = ExprTyper::new(&mut Environment::new(
-            &mut uid,
+            ids,
             &["somemod".to_string()],
             &modules,
             &mut vec![],