+++ CHANGELOG.md
@@ -2,8 +2,13 @@
 
 ## Unreleased
 
+- A WebAssembly version of the compile is now available for use in JavaScript
+  and other WebAssembly environments.
 - New projects include Hex badges and a link to Hexdocs.
 - Enhance type mismatch errors in the presence of try.
+- Enhance type mismatch error for an inconsistent try.
+- Fixed a bug where sometimes type variable could be reused result in incorrect
+  non-deterministic type errors.
 
 ## v0.19.0 - 2022-01-12